M1171-M1200: compile verification + test execution + bug fixes
- 39/39 .cpp files compile (27 tests + 12 tools)
- 27/27 tests PASS from src/joinrenum/ working directory
- Fixed 10 type/path bugs:
  - test_count_oracle.cpp: printf %d → %zu for size_t dim()
  - test_count_oracle_full.cpp: fprintf %d → %zu for size_t dim()
  - test_enumerator.cpp: printf %lld → %d for int counters (6 places)
  - test_join_tree.cpp: printf %d → %lld for long long (2 places)
  - test_join_triangle.cpp: Ra.tbl → Ra.csv (file doesn't exist)
  - test_join_baseline_full.cpp: added .csv fallback paths
- Verified AJB breakpoint system: STATE/TIMER/BP/DEBUG/RESULTS/PROBE/MEM/TRACE

+++ src/joinrenum/tests/test_count_oracle.cpp
@@ -152,8 +152,8 @@ int main(int argc, char* argv[]) {
         }
         printf("[AJB] Generated %d synthetic %d-D points (data.txt fallback)\n", synth_n, synth_dim);
     }
-    printf("[AJB] Loaded %zu points, dim=%d\n",
-           points.size(), points.empty() ? 0 : points[0].dim());
+    printf("[AJB] Loaded %zu points, dim=%zu\n",
+           points.size(), points.empty() ? (size_t)0 : points[0].dim());
 
     AJBMemSnapshot mem_after_load;
     mem_after_load.Capture();
